requirements.txt
Django==2.1.2

# For use on settings.py
dj-database-url==0.5.0
python-decouple==3.1

# Django Rest Framework
djangorestframework==3.8.2

# Coreapi needed for schema support do DRF.
coreapi==2.3.3

# A JSON Web Token authentication plugin for the Django REST Framework.
djangorestframework-simplejwt==3.2.3

# For DRF filters
django-filter==2.0.0

# Custom Admin layout
django-admin-interface==0.8.0

# Collection of custom extensions for the Django. Management commands, additional database fields, admin extensions
django-extensions==2.1.3

# IPython (Python Console)
ipython==7.0.1

# To AWS S3
django-storages==1.7.1  # provide a variety of storage backends in a single library.
boto3==1.9.23           # Amazon Web Services (AWS) SDK for Python

# Needed for Heroku
gunicorn==19.9.0  # Python WSGI HTTP Server for UNIX
psycopg2==2.7.5   # PostgreSQL adapter for the Python

###################################
